--story=887060303 环境管理服务对接上云版作业平台
feat: 添加agent状态查询逻辑，添加节点时 向T_NODE表写入agent版本

+++ src/backend/ci/core/environment/api-environment/src/main/kotlin/com/tencent/devops/environment/pojo/job/agentreq/QueryAgentStatusFromNodemanReq.kt
@@ -31,7 +31,7 @@ import io.swagger.annotations.ApiModelProperty
 
 data class QueryAgentStatusFromNodemanReq(
     @ApiModelProperty(value = "主机ID")
-    val bkHostId: List<Int>?,
+    val bkHostId: List<Long>?,
     @ApiModelProperty(value = "搜索条件")
     val conditions: List<Condition<List<String>>>?,
     @ApiModelProperty(value = "展示额外信息")
